requirements.txt
fastapi
uvicorn
pandas
matplotlib
seaborn
python-dotenv
requests
python-multipart
jinja2
numpy
scipy
aiofiles
